fsmonitor--daemon: background daemon must free the console on windows
Teach "git fsmonitor--daemon run" to call FreeConsole() when started
in the background by "git fsmonitor--daemon start" on Windows.

The background process was holding a handle to the inherited Win32
console despite being passed stdin/out/err set to /dev/null.  This
caused command prompts and powershell terminal windows to hang in
"exit" waiting for the last console handle to be released.

(This problem was not seen in git-bash type terminal windows because
they don't have a Win32 console attached to them.)

Signed-off-by: Jeff Hostetler <jeffhost@microsoft.com>